+++ ports/csharp/src/gui/widgets/BcsWidget.cs
@@ -11,26 +11,16 @@ namespace Bcs.Gui.Widgets
         private int _height = 0;
         private int _x = 0;
         private int _y = 0;
+        private bool _hasFocus = false;
         private readonly object _lock = new object();
 
         public BcsWidget(BcsWidget parent = null) : base(parent)
         {
         }
 
-        public void Show()
-        {
-            lock (_lock) { _visible = true; }
-        }
-
-        public void Hide()
-        {
-            lock (_lock) { _visible = false; }
-        }
-
-        public bool IsVisible
-        {
-            get { lock (_lock) { return _visible; } }
-        }
+        public void Show() { lock (_lock) { _visible = true; } }
+        public void Hide() { lock (_lock) { _visible = false; } }
+        public bool IsVisible { get { lock (_lock) { return _visible; } } }
 
         public void Resize(int width, int height)
         {
@@ -55,5 +45,41 @@ namespace Bcs.Gui.Widgets
             get { lock (_lock) { return _enabled; } }
             set { lock (_lock) { _enabled = value; } }
         }
+
+        public bool HasFocus
+        {
+            get { lock (_lock) { return _hasFocus; } }
+            set { lock (_lock) { _hasFocus = value; } }
+        }
+
+        public override bool Event(BcsEvent e)
+        {
+            switch (e.Type)
+            {
+                case EventType.MouseButtonPress:
+                    MousePressEvent(e);
+                    return true;
+                case EventType.MouseButtonRelease:
+                    MouseReleaseEvent(e);
+                    return true;
+                case EventType.KeyPress:
+                    KeyPressEvent(e);
+                    return true;
+                case EventType.KeyRelease:
+                    KeyReleaseEvent(e);
+                    return true;
+                case EventType.Paint:
+                    PaintEvent(e);
+                    return true;
+                default:
+                    return base.Event(e);
+            }
+        }
+
+        protected virtual void MousePressEvent(BcsEvent e) {}
+        protected virtual void MouseReleaseEvent(BcsEvent e) {}
+        protected virtual void KeyPressEvent(BcsEvent e) {}
+        protected virtual void KeyReleaseEvent(BcsEvent e) {}
+        protected virtual void PaintEvent(BcsEvent e) {}
     }
 }