+++ ports/rust/src/core/kernel/mod.rs
@@ -1,0 +1,5 @@
+pub mod bcs_event;
+pub mod bcs_event_dispatcher;
+pub mod bcs_object;
+pub mod bcs_event_loop;
+pub mod bcs_kernel;