+++ ports/rust/src/gui/kernel/mod.rs
@@ -1,0 +1,1 @@
+pub mod bcs_input_arbitrator;